fix(types): Use generics for ListItem component's items prop typing instead of 'any', and prep add rework ShoppingList type
Use generics to ensure type safety between component's props

Rework the ShoppingList type by adding a 'position' prop to keep track of the list's position in a list. Also, rename this type to 'BaseProps' since it become the base for the ShoppingListItem type as well, since it will also need to keep track of its position in lists.